consolidate code + working on array limit
