+++ src/constants.ts
@@ -1,0 +1,2 @@
+
+export const indentation = '  '; // TODO: Update based off the config (spacing/tab amount)
